style(rubocop): change block style to be consistent (#8)
We currently allow any style of block, which leads to code such as:

```
arr.select { |a|
  a == true
}
```

This occurs the most in specs:

```
let(:customer) {
  create(:customer)
}
```

This is difficult to read because it is inconsistent. When enforced as line_count (default) it would read:

```
arr.select { |a| a == true }

let(:customer) { create(:customer) }
```

And only breaks to do/end blocks when the code is too complex to fit on a single line.